pyproject.toml
[project]
name = "moneyexchange-api"
version = "0.1.0"
description = "MoneyExchange API is a Django-based application providing currency exchange rates, user balance management, and transaction history. The API integrates with an external ExchangeRate API, uses JWT authentication, and is containerized using Docker and managed via Poetry."
authors = [
    {name = "Mike Shmygel",email = "m.shmygel@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "django-debug-toolbar (>=5.0.1,<6.0.0)",
    "django-jazzmin (>=3.0.1,<4.0.0)",
    "drf-yasg (>=1.21.10,<2.0.0)",
    "pytest (>=8.3.5,<9.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-django = "^4.10.0"
coverage = "^7.7.0"


[tool.pytest.ini_options]
python_files = [
    "test_*.py",
    "*_test.py"
]
DJANGO_SETTINGS_MODULE = "moneyexchange_project.settings"

